fix(ai): fix PMD warning in OpenAiLlmClient
Use long type directly for TokenUsage constructor instead of
unnecessary intValue() calls. TokenUsage already uses long type.
Suppress deprecation warning for getGenerationTokens() which
is deprecated in Spring AI 1.0.0-M6.

+++ ai-impl/ai-llm/src/main/java/io/github/afgprojects/framework/ai/llm/openai/OpenAiLlmClient.java
@@ -170,9 +170,9 @@ public class OpenAiLlmClient implements LlmClient {
         if (metadata != null && metadata.getUsage() != null) {
             var usage = metadata.getUsage();
             tokenUsage = new TokenUsage(
-                    usage.getPromptTokens().intValue(),
-                    usage.getGenerationTokens().intValue(),
-                    usage.getTotalTokens().intValue()
+                    usage.getPromptTokens(),
+                    usage.getGenerationTokens(),
+                    usage.getTotalTokens()
             );
         }
